consider blockquotes as a blocks instead of lines
This is roughly the same system as for preformatted blocks.

Even though the spec implies that a line starting with ">" is a
blockquote type line and should be a quote on its own, it is so
frequently used as a block-like type that taking the spec literally
causes rendering issues on many capsules. To be clear, we get blocks
like this:

> Lorem ipsum dolor sit amet,
> consectetur adipiscing elit.
>
> Sed do eiusmod tempor incididunt
> ut labore et dolore magna aliqua.

where we should instead have something like:

> Lorem ipsum dolor sit amet, consectetur adipiscing elit.
> Sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.

To the people's defense, it is usual to quote a message using the first
syntax. As this issue is minor and the spec is not very adamant on
things here, I don't think there is much wrongdoing in treating the
first syntax as the norm, but the code is open to changes!